0.2.0: update for Rust 1.0.0
The crate no longer depends on any gated features, but can be used with
the unstable `no_std` with the `no_std` crate feature.

I have decided that memory safety is important and that aliasing, even
if the code that would lead to it mostly doesn’t make sense, should be
truly forbidden. Fixes #9.