0.3.0: update MuCell mostly in line with RefCell.
[breaking-change]: `mucell_ref_type!` is gone.

This is backwards-compatible with 0.2.0 except for the
`mucell_ref_type!` macro (and surrounding convention) which is now
obsolete. Users of the `mucell_ref_type!` macro should refactor their
code to use `Ref::map` instead.

Given a `MuCell<Foo>`, you would have written and used your reference
type like this:

```rust
mucell_ref_type! {
    struct BarRef<'a>(Foo),
    impl Deref -> _,
    data: Bar = |x| x.bar()
}

fn bar<'a>(foo: &'a MuCell<Foo>) -> BarRef<'a, Foo> {
    // ...
    BarRef::from(foo)
}
```

This should now be done like this:

```rust
use mucell::Ref;

fn bar<'a>(foo: &'a MuCell<Foo>) -> Ref<'a, Bar> {
    // ...
    Ref::map(foo.borrow(), |x| x.bar())
}
```

This new approach is macro-free, simpler, clearer and more extensible.
I’m looking forward to the `std::cell` versions of these functions,
which are behind the `cell_extras` feature gate, becoming stable!

(My `Ref` is more powerful than `std::cell::Ref`, which can only contain
a reference; my `Ref<'a, T>` is thus equivalent to its `Ref<'a, &'a T>`;
I am producing an RFC to fix that backwards-compatibly upstream.)

See also `Ref::filter_map` for functions which return an `Option`;
this is something which was sometimes necessary and which the macro
approach had not handled.