0.3.5: update for Rust no_std compatibility.
`#![no_std]` is now stable and when used core is imported automatically
(it’s stable too).

Actually *using* the `no_std` feature still requires nightly, because of
the lang items involved. But the library itself will build with `no_std`
on stable, for all the good it’ll do you.